fix: move node update event to pov
