+++ ledger/cache.go
@@ -10,11 +10,12 @@ import (
 	"time"
 
 	"github.com/bluele/gcache"
+	"go.uber.org/zap"
+
 	"github.com/qlcchain/go-qlc/common/storage"
 	"github.com/qlcchain/go-qlc/common/types"
 	"github.com/qlcchain/go-qlc/ledger/relation"
 	"github.com/qlcchain/go-qlc/log"
-	"go.uber.org/zap"
 )
 
 type MemoryCache struct {
